// frontend/src/api/models/RuleTypeEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * * `transaction_amount` - Transaction Amount
 * * `velocity` - Velocity
 * * `geographic` - Geographic
 * * `time_based` - Time Based
 * * `account_activity` - Account Activity
 */
export enum RuleTypeEnum {
    TRANSACTION_AMOUNT = 'transaction_amount',
    VELOCITY = 'velocity',
    GEOGRAPHIC = 'geographic',
    TIME_BASED = 'time_based',
    ACCOUNT_ACTIVITY = 'account_activity',
}